requirements.txt
aiohttp==3.7.4
aioredis==1.3.1
Boruta
catboost==1.0.4
click==7.1.2
cloudpathlib==0.6.2
fsspec==2022.11.0
gcsfs==2022.11.0
google-cloud-storage==2.3.0
hyperopt== 0.2.5
lightgbm==3.2.1
matplotlib==3.5.1
mkdocs
mkdocs-git-revision-date-plugin
mkdocs-material
mkdocstrings
neptune==1.2.0
numpy==1.19.5
pandas==1.1.4
pillow==8.3.2
plotly==5.3.1
protobuf==3.20.0
pylint==2.7.2
pymdown-extensions
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
pytest-order==1.0.0
pytest-parallel==0.1.1
pytest-xdist==2.4.0
python-dotenv==0.19.0
ray[default]==1.4.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn
shap
sklearn-relief==1.0.0b2
tqdm==4.60.0
typing-extensions==4.4.0
xgboost==1.4.2
xlrd==1.2.0
